Use a python virtual env and a non root user setup
Get rid of the warning about using a python install as root as
described in #147.

That resulted in a few changes in `Dockerfile.api`. Mainly the instructions
for docker where split into two parts.
Those, that root has to perform and
those, that can be done as a normal user.

In Openshift the user at runtime will not be root, but a random uid with
a group-id of root. Therefore we choose
a pseudo random uid of 1001 and give add that user to the root group.

To be able to use the virtual env in UWSGI, we need to set another ENV
variable to point to the venv.